Add files via upload
Analyzed a study using SAS to evaluate "Drug A" effects on fasting blood sugar in 3,400 patients across low-dose, high-dose, and placebo groups. After 10 weeks, a mean difference of 12.4 was found between the low and high-dose groups. ANOVA with a p-value of 0.005 rejected the null hypothesis. The Scheffe test revealed the most significant difference between the low-dose and placebo groups, highlighting Drug A’s potential impact on fasting blood sugar levels.